src/requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.9
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
psycopg==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentence-transformers==3.3.1
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.21.0
